[Audio] avoid overflow
git-svn-id: svn://svn.berlios.de/avidemux/branches/avidemux_2.6_branch_mean@4704 30592fff-cfd4-0310-bf0e-8235a3c7be26

+++ avidemux/ADM_coreMuxer/include/ADM_audioClock.h
@@ -10,7 +10,7 @@ class audioClock
 {
         protected:
                 uint32_t        _frequency;
-                uint32_t        _nbSamples;
+                uint64_t        _nbSamples;
                 uint64_t        _baseClock;
 
         public:
